+++ packages/apm/src/hubextensions.ts
@@ -60,9 +60,13 @@ function startSpan(spanOrSpanContext?: Span | SpanContext, forceNoChild: boolean
     span.sampled = Math.random() < sampleRate;
   }
 
-  if (span.sampled) {
-    const experimentsOptions = (client && client.getOptions()._experiments) || {};
-    span.initFinishedSpans(experimentsOptions.maxSpans as number);
+  // We always want to record spans independent from the sample rate
+  const experimentsOptions = (client && client.getOptions()._experiments) || {};
+  span.initFinishedSpans(experimentsOptions.maxSpans as number);
+
+  // If we do not have an op by now by default we set a name otherwise the server will discard the transaction/span
+  if (span.op === undefined) {
+    span.op = 'op';
   }
 
   return span;